Moved tests to subfolder
