+++ tests/e2e/e2e_test.go
@@ -9,13 +9,13 @@ import (
 	"os/exec"
 	"testing"
 
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/key"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/network"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/packageman"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/root"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/subnet"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/subnet/local"
-	_ "github.com/ava-labs/avalanche-cli/tests/e2e/subnet/public"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/key"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/network"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/packageman"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/root"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/subnet"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/subnet/local"
+	_ "github.com/ava-labs/avalanche-cli/tests/e2e/testcases/subnet/public"
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	"github.com/onsi/gomega/format"
