+++ tests/e2e/testcases/packageman/suite.go
